% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.character.treeprof}
\alias{as.character.treeprof}
\alias{as.character.treeprof_fun_table}
\alias{as.character.treeprof_summary,}
\title{Transform treeprof Output Into Text Representation}
\usage{
\method{as.character}{treeprof}(x, mode = "TEXT", id.start = 1L,
  depth = 10L, disp.thresh = 5L, disp.unit = "auto")
}
\arguments{
\item{x}{treeprof object}

\item{mode}{1 length character either TEXT or HTML}

\item{id.start}{int what level to display from}

\item{depth}{int 1 length how many levels to show}

\item{disp.thresh}{number of mils below which a branch isn't shown}

\item{disp.unit}{character(1L) use "mills" to show things really to 1000 total
time, or a time unit (e.g. "seconds", "microseconds"); default will find
time unit that displays most sensibly}
}
\value{
character vector with as many elements as there are lines of output
}
\description{
Transform treeprof Output Into Text Representation
}

