% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.table.treeprof}
\alias{as.data.table.treeprof}
\alias{as.data.table.treeprof_fun_table}
\title{Convert a Treeprof object to `data.table`}
\usage{
\method{as.data.table}{treeprof}(x, keep.rownames = FALSE)
}
\arguments{
\item{x}{a \code{`treeprof`} object}
}
\value{
a \code{`data.table`} object
}
\description{
Primarily used to display the underlying \code{`data.table`} structure of a
\code{`treeprof`} object
}

