% Generated by roxygen2 (4.0.1): do not edit by hand
\name{by_fun}
\alias{by_fun}
\title{Per Function Summary}
\usage{
by_fun(x)
}
\arguments{
\item{x}{treeprof object}
}
\value{
treeprof_fun_table object, which inherits from \code{`data.table`}
}
\description{
Computes execution times for each function, accounting for any repeats, but
ignoring contrbutions from nested repeats when summing up total times.
}

