% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_message}
\alias{clean_message}
\title{Outputs a Message For Overwriting}
\usage{
clean_message(msg, verbose)
}
\arguments{
\item{msg}{character}

\item{verbose}{logical(1L)}
}
\description{
Outputs a Message For Overwriting
}
\keyword{internal}

