% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collapse_passthru_fun}
\alias{collapse_passthru_fun}
\title{Remove All Intermediate Levels of a Passthrough}
\usage{
collapse_passthru_fun(x, passthru)
}
\arguments{
\item{x}{a treeprof object}

\item{passthru}{a \code{`passthru_fun`} object}
}
\value{
a modified treeprof object
}
\description{
Motivated by functions like \code{`\link{try}`} and \code{`\link{tryCatch}`}
that have multiple layers of internal calls but then eventually go to
evaluate user code.  The idea here is to collapse all those intermediate
levels so they don't clutter the profile tree.
}
\details{
Most of these functions seem to have a common pattern: there is the function
itself that is called, and then at some point, one final function is called
before the user code is evaluated.  We capture this by looking for the initial
function, and then within the children of that function, look for the final
function subject to the condition that the final function is not followed by
a disallowed function.  This last condition is necessary because some functions
like \code{`\link{tryCatch}`} call themselves in order to register multiple
handlers so we would otherwise think we have exited the function when in
reality we haven't.

Clearly this isn't completely fool proof, especially if someone writes a
function with the same name as some of these base passthru functions.
}

