% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collapse_passthru_funs}
\alias{collapse_passthru_funs}
\title{Applies All Defined Passthrus}
\usage{
collapse_passthru_funs(x, passthrus = passthru_defined)
}
\arguments{
\item{x}{a \code{`treeprof`} object}

\item{passthrus}{a list of \code{`passthru_fun`} objects}
}
\value{
a \code{`treeprof`} object
}
\description{
Applies All Defined Passthrus
}
\seealso{
\code{`\link{collapse_passthru_fun}`}, \code{`\link{passthru_fun}`}
}
\keyword{internal}

