% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_descendant_nodes}
\alias{get_child_nodes}
\alias{get_descendant_nodes}
\title{Get \code{`ids`} of Descendant Nodes}
\usage{
get_descendant_nodes(x, id.par)
}
\arguments{
\item{x}{a \code{`treeprof`} object}

\item{id.par}{the id of the node to get the children of}
}
\value{
one column data.table containing requested ids (for joining with)
  `x` if desired, note this will also include the parent node
}
\description{
\code{`get_descendant_nodes`} gets all nodes that have for direct or indirect
parent \code{`id.par`}, including \code{`id.par`}, whereas \code{`get_child_nodes`}
only returns the direct children (and also the parent).
}
\keyword{internal}

