% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_par_nodes}
\alias{get_par_nodes}
\title{Get \code{`ids`} of Parent Node For Each Node}
\usage{
get_par_nodes(x)
}
\arguments{
\item{x}{a \code{`treeprof`} object}
}
\value{
integer vector where the first item has the parent id of node id 1, the
  second of node id 2, etc.  Nodes that don't exist in the input will have 0L
  for parent (root node will too)
}
\description{
Get \code{`ids`} of Parent Node For Each Node
}
\keyword{internal}

