% Generated by roxygen2 (4.0.1): do not edit by hand
\name{melt_prof}
\alias{melt_prof}
\title{Generate The Final Treeprof Object}
\usage{
melt_prof(mx, levels.to.exclude)
}
\arguments{
\item{mx}{a character matrix where each row represents a tick dump, and each
column a level in the stack}
}
\value{
a \code{`treeprof`} object
}
\description{
Generate The Final Treeprof Object
}
\keyword{internal}

