% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize}
\alias{normalize}
\title{Converts to desired time units}
\usage{
normalize(x, disp.unit)
}
\arguments{
\item{treeprof}{data.table object}

\item{disp.unit}{what the desired display unit is}
}
\value{
a treeprof_norm object with added normalized time columns
}
\description{
Will add two columns to the underlying \code{`data.table`} object with the
normalized \code{`n`} and \code{`n.self`} values.  Additionally will attach
an attribute to the return object
}
\keyword{internal}

