% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_lines}
\alias{parse_lines}
\title{Converts \code{`\link{Rprof}`} Output to Matrix}
\usage{
parse_lines(file)
}
\arguments{
\item{a}{file name}
}
\value{
a matrix containing the parsed and reversed call stacks, with one
  additional NA column added at the end
}
\description{
Converts \code{`\link{Rprof}`} Output to Matrix
}
\keyword{internal}

