% Generated by roxygen2 (4.0.1): do not edit by hand
\name{passthru_fun}
\alias{passthru_fun}
\title{Represents Pass Through Function We Wish to Condense In Call Stack}
\usage{
passthru_fun(enter, exit, dont.exit.if)
}
\arguments{
\item{enter}{character 1 length the name of the pass through function}

\item{exit}{character 1 length the name of the function that then evaluates
the rest of the code}

\item{depth}{integer 1 length the distance between enter and exit in terms of
levels in the call stack; this is necessary to deal with potentially nested
pass through functions; note this includes the entry and exit levels, so if
the entry and exit levels are 1 apart, the depth should be 2, not 1}
}
\value{
a passthru_fun class S3 object
}
\description{
Prime example is `try`, which generates a whole bunch of calls that are not
really useful in trying to figure out what's going on in the profile.
}
\details{
These are referred to as pass through functions because they wrap around
other code, so we're interested in seeing what the rest of the code is doing
}

