% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_rprof}
\alias{run_rprof}
\title{Actually Run the Code and Capture \code{`Rprof`} Output}
\usage{
run_rprof(expr.quoted, interval, target.time, times, file,
<<<<<<< HEAD
<<<<<<< HEAD
  frame = parent.frame(), gc.torture, verbose)
=======
  frame = parent.frame(), gc.torture)
>>>>>>> 75d32eb92dbce607391b8bcf59be3a129015cdd0
=======
  frame = parent.frame(), gc.torture, verbose)
>>>>>>> 2a504db3379c01ea7079dc7c05a7fc2800124694
}
\arguments{
\item{expr.quoted}{a quoted expression}

\item{frame}{a frame to evaluate the quoted expression in}
}
\value{
list containing the file name of the Rprof data plus some meta data
}
\description{
Actually Run the Code and Capture \code{`Rprof`} Output
}
\keyword{internal}

