% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shinyfy}
\alias{shinyfy}
\title{Interactive UI For Exploring Benchmarks}
\usage{
shinyfy(x, id.start = 1L, depth = 10L)

shinyfy(x, id.start = 1L, depth = 10L)
}
\arguments{
\item{x}{a treeprof object}

\item{id.start}{integer(1L) what node to start displaying from}

\item{depth}{integer(1L) how deep down the tree to start off showing}

\item{x}{a treeprof object}

\item{id.start}{integer(1L) what node to start displaying from}

\item{depth}{integer(1L) how deep down the tree to start off showing}
}
\description{
Creates an interactive Shiny environment for users to browse through the
\code{`Rprof`} results in tree structure.

Creates an interactive Shiny environment for users to browse through the
\code{`Rprof`} results in tree structure.
}
\note{
ideally this would be implemented as an S3 method to the \code{`shine`}
  generic, but Shiny does not currenlty define such a method so we're stuck
  with this.

ideally this would be implemented as an S3 method to the \code{`shine`}
  generic, but Shiny does not currenlty define such a method so we're stuck
  with this.
}
\examples{
\dontrun{
x <- treeprof(data.frame(a=1:10, b=letters[1:10]))
shinify(x)
}
\dontrun{
x <- treeprof(data.frame(a=1:10, b=letters[1:10]))
shinify(x)
}
}

