% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sort.treeprof}
\alias{sort.treeprof}
\title{Sort A Treeprof}
\usage{
\method{sort}{treeprof}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{treeprof object}

\item{decreasing}{whether to sort descending or not}

\item{treeprof}{object, "sorted"}
}
\description{
Each branch is sorted in turn recursively.  The treeprof object is re-id'd
in the order of the sort.
}
\note{
set \code{`decreasing=TRUE`} as otherwise the tree stops making sense;
  default is set to \code{`FALSE`} for compatibility with generic.
}

