% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize}
\alias{summarize}
\title{Compute High Level Summary of \code{`treeprof`}}
\usage{
summarize(x)
}
\arguments{
\item{x}{\code{`treeprof_norm`} object}
}
\value{
\code{`treeprof_summary`} object
}
\description{
Compute High Level Summary of \code{`treeprof`}
}
\keyword{internal}

