% Generated by roxygen2 (4.0.1): do not edit by hand
\name{time_format}
\alias{time_format}
\alias{time_scale}
\title{Display Time In Reasonable Units}
\usage{
time_format(x, signif = 4L, inc.units = TRUE, trim = TRUE)
}
\arguments{
\item{x}{numeric to format in seconds}

\item{scale}{how to compute time units, "auto" or "kilosecond" - "femtosecond"}

\item{signif}{integer(1L) how many significant digits the largest number should have}

\item{inc.units}{logical(1L) whether to paste the time units to each number}

\item{trim}{logical(1L) set to FALSE to keep leading whitespace to align all
numbers}
}
\value{
character formatted numbers for \code{`time_format`}, a
  \code{`time_scaled`} S3 object for \code{`time_scale`} containing scaled
  times and units
}
\description{
Formats such that \code{`max(min(x), .005 * max(x))`} is at worst one decimal
place.  This should work for multi value vectors as well as single value
ones.
}
\keyword{internal}

