% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trim_branch}
\alias{trim_branch}
\alias{trim_branch_fast}
\title{Presents A Pruned Version of A Branch}
\usage{
trim_branch(x, i, depth = 5L)
}
\arguments{
\item{x}{data.table}

\item{i}{the node id}

\item{depth}{how many levels below \code{`i`} to show}

\item{disp.thresh}{nodes below which speed threshold to keep (5 mills default)}
}
\value{
data.table pruned as described
}
\description{
Depending on the function, keeps target node (\code{`i`}) children to depth
\code{`depth`}, as well as target node parent's, and those parent's direct
children (uncles, etc.), or trims all nodes that execute faster than a
threshold time.
}
\details{
Typically \code{`trim_branch_fast`} is applied after \code{`trim_branch`} as
otherwise \code{`trim_branch`} would show the branch all the way to the
leaves.
}
\keyword{internal}

